Basket tab created

+++ src/Entity/Product.java
@@ -4,7 +4,7 @@ public class Product {
     private int id;
     private String name;
     private String code;
-    private double price;
+    private int price;
     private int stock;
 
     public Product() {
@@ -34,11 +34,11 @@ public class Product {
         this.code = code;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
